refactor: renamed parent "User" folder to "user"
