+++ server/src/main/java/com/app/server/authentication/JwtService.java
@@ -1,0 +1,5 @@
+package com.app.server.authentication;
+
+public class JwtService {
+
+}
