+++ server/src/main/java/com/app/server/user/UserController.java
@@ -1,4 +1,4 @@
-package com.app.server.controllers;
+package com.app.server.user;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
